+++ src/scalar.hpp
@@ -3,7 +3,7 @@
 #include "util/slice.hpp"
 #include "util/pyobj.hpp"
 #include "util/json/json.hpp"
-#include "util/tsv.hpp"
+#include "util/xsv.hpp"
 
 namespace ss{
